moved base_schema methods into utils, adjust all function calls; still need to setup actual RecordingExtractor conversion
